Harden Invoker test reliability (8): Extract shared TaskRunner test fixture
Review claim: Extract Function moves `createAutoCompleteExecutor` into one shared test helper with behavior unchanged.
Review lane: refactor
Review unit: ownership-refactor
Safety invariant: Assertions and production code remain unchanged, and the complete execution-engine package test stays green.
Slice rationale: One top-level symbol move establishes the seam for later one-cluster-per-workflow decomposition.
Architectural effect: Two TaskRunner suites depend on one fixture owner.
Goal: Remove duplicated fixture ownership without moving test clusters.
Motivation: The same helper exists in two multi-thousand-line suites.
Alternative considerations: A bulk decomposition is not locally reviewable.
Implementation details: Create one helper module, move exactly the duplicated function, and repoint both callers.
Non-goals: No assertion changes, production edits, extra symbol moves, Orchestrator moves, or skipped-test work.
Layer: app_regression
Layer exception: allowed
Layer exception rationale: This lower-layer refactor intentionally follows the higher-layer reliability work so structural movement cannot conceal behavior or proof changes.
Feature state: active
Files:
- packages/execution-engine/src/__tests__/helpers/task-runner-fixtures.ts
- packages/execution-engine/src/__tests__/task-runner.test.ts
- packages/execution-engine/src/__tests__/task-runner-fix-publish-and-ssh.test.ts
Change types:
- packages/execution-engine/src/__tests__/helpers/task-runner-fixtures.ts: create
- packages/execution-engine/src/__tests__/task-runner.test.ts: test-only
- packages/execution-engine/src/__tests__/task-runner-fix-publish-and-ssh.test.ts: test-only
Acceptance criteria:
- Exactly one live `createAutoCompleteExecutor` implementation remains.
- Assertions and production files are unchanged.
- `cd packages/execution-engine && pnpm test` exits 0.
- Final `pnpm test` exits 0.

Context:
  wf-1787375663674-18/schedule-linux-visual-proof (c273b06): Review claim: Scheduled CI policy runs Linux visual proof and retains failure artifacts without making it merge-required.
Review lane: policy
Review unit: tooling-policy
Safety invariant: Linux pixel failures affect scheduled runs only; merge-required conditions remain unchanged.
Slice rationale: Scheduling follows the independently proven screenshot behavior.
Architectural effect: Scheduled CI becomes the owner of recurring Linux screenshot execution.
Goal: Run Linux pixel proof automatically each night.
Motivation: Local-only screenshot proof cannot catch regressions continuously.
Alternative considerations: Immediate merge gating risks queue blockage from platform noise.
Implementation details: Extend the existing Playwright wrapper and scheduled job to set the proof flag, run the visual spec, and retain diffs on failure.
Non-goals: No screenshot implementation changes and no merge-required condition changes.
Layer: e2e_regression
Feature state: active
Files:
- scripts/test-suites/optional/40-playwright-app.sh
- .github/workflows/ci.yml
Change types:
- scripts/test-suites/optional/40-playwright-app.sh: modify
- .github/workflows/ci.yml: config-only
Acceptance criteria:
- Scheduled/workflow_dispatch CI sets the visual proof flag.
- Failure diffs are retained.
- Merge-required conditions remain unchanged.
- `pnpm test` exits 0.

Solution:
  Review claim: Extract Function moves `createAutoCompleteExecutor` into one shared test helper with behavior unchanged.
Review lane: refactor
Review unit: ownership-refactor
Safety invariant: Assertions and production code remain unchanged, and the complete execution-engine package test stays green.
Slice rationale: One top-level symbol move establishes the seam for later one-cluster-per-workflow decomposition.
Architectural effect: Two TaskRunner suites depend on one fixture owner.
Goal: Remove duplicated fixture ownership without moving test clusters.
Motivation: The same helper exists in two multi-thousand-line suites.
Alternative considerations: A bulk decomposition is not locally reviewable.
Implementation details: Create one helper module, move exactly the duplicated function, and repoint both callers.
Non-goals: No assertion changes, production edits, extra symbol moves, Orchestrator moves, or skipped-test work.
Layer: app_regression
Layer exception: allowed
Layer exception rationale: This lower-layer refactor intentionally follows the higher-layer reliability work so structural movement cannot conceal behavior or proof changes.
Feature state: active
Files:
- packages/execution-engine/src/__tests__/helpers/task-runner-fixtures.ts
- packages/execution-engine/src/__tests__/task-runner.test.ts
- packages/execution-engine/src/__tests__/task-runner-fix-publish-and-ssh.test.ts
Change types:
- packages/execution-engine/src/__tests__/helpers/task-runner-fixtures.ts: create
- packages/execution-engine/src/__tests__/task-runner.test.ts: test-only
- packages/execution-engine/src/__tests__/task-runner-fix-publish-and-ssh.test.ts: test-only
Acceptance criteria:
- Exactly one live `createAutoCompleteExecutor` implementation remains.
- Assertions and production files are unchanged.
- `cd packages/execution-engine && pnpm test` exits 0.
- Final `pnpm test` exits 0.

Invoker-Finalize-Id: c4cef65d-416d-49bd-b04d-3a165a65c78a
